refactor(frontend): update menu dialogs
